Change all createentity args to be ints
The script command createentity() is always an int. But not only that,
every time createentity() is used, its arguments are always treated like
ints. Always. I knew that vx/vy were floats because of the int casts in
the function, but I didn't even realize that xp/yp were floats, too,
until I checked just now! That's how much they're treated like ints.

All int casts in createentity() have also been removed, due to being
unnecessary (either because of us suppressing MSVC implicit conversion
warnings, or because there are now no longer any conversions happening).

+++ desktop_version/src/Entity.h
@@ -74,8 +74,8 @@ public:
 
     int crewcolour(int t);
 
-    void createentity(float xp, float yp, int t, float vx = 0, float vy = 0,
-                      int p1 = 0, int p2 = 0, int p3 = 320, int p4 = 240 );
+    void createentity(int xp, int yp, int t, int vx = 0, int vy = 0,
+                      int p1 = 0, int p2 = 0, int p3 = 320, int p4 = 240);
 
     bool updateentities(int i);
